Add first text component; update some existing usages
Added a basic TextHeading1 component which represents an H1 (currently
red) and takes in a value. Added a shared file for fonts and put a
default font family in there. Added an index for exporting the text
headings. Added white and gray to the shared colors. Updated
SubmitButton to use the white defined by colors. Updated app.js to use
the gradient defined in colors. Added a text heading. Gave an onPress
function to the submit button just for fun.